add forground service and fix permission issue
