+++ app/build.gradle
@@ -7,11 +7,11 @@ plugins {
 }
 
 android {
-    namespace 'com.example.locationtracker'
+    namespace 'dev.nura.locationtracker'
     compileSdk 34
 
     defaultConfig {
-        applicationId "com.example.locationtracker"
+        applicationId "dev.nura.locationtracker"
         minSdk 26
         targetSdk 34
         versionCode 1
@@ -58,6 +58,8 @@ dependencies {
     implementation libs.library.base
     implementation libs.androidx.security.crypto
     implementation libs.play.services.location
+    implementation libs.play.services.maps
+    implementation libs.google.maps.services
 
     testImplementation libs.junit
     androidTestImplementation libs.androidx.junit
