--- app/src/main/java/com/example/locationtracker/Utils/Utils.kt
@@ -1,20 +1,0 @@
-package com.example.locationtracker.Utils
-
-import android.view.Window
-import androidx.core.view.WindowCompat
-import androidx.core.view.WindowInsetsCompat
-import androidx.core.view.WindowInsetsControllerCompat
-
-fun Window.hideSystemUI() {
-    WindowCompat.setDecorFitsSystemWindows(this, false)
-    WindowInsetsControllerCompat(this, this.decorView).let { controller ->
-        controller.hide(WindowInsetsCompat.Type.systemBars())
-        controller.systemBarsBehavior =
-            WindowInsetsControllerCompat.BEHAVIOR_DEFAULT
-    }
-}
-
-fun String.isValidEmail(): Boolean {
-    val emailRegex = Regex("[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}")
-    return emailRegex.matches(this)
-}
