--- app/src/main/java/com/example/locationtracker/ui/MapViewFragment.kt
@@ -1,34 +1,0 @@
-package com.example.locationtracker.ui
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.locationtracker.R
-import com.example.locationtracker.databinding.FragmentMapViewBinding
-
-
-class MapViewFragment : Fragment() {
-
-    private var _binding: FragmentMapViewBinding? = null
-    private val binding get() = _binding!!
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentMapViewBinding.inflate(layoutInflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-
-    companion object {
-        const val TAG: String = "MapViewFragment"
-    }
-
-}
