--- app/src/main/java/com/example/locationtracker/ui/SplashFragment.kt
@@ -1,58 +1,0 @@
-package com.example.locationtracker.ui
-
-import android.os.Bundle
-import android.os.CountDownTimer
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.navigation.fragment.findNavController
-import com.example.locationtracker.R
-import com.example.locationtracker.commen.AppPreferences
-import com.example.locationtracker.databinding.FragmentSplashBinding
-
-
-class SplashFragment : Fragment() {
-
-    private var _binding: FragmentSplashBinding? = null
-    private val binding get() = _binding!!
-
-    private val timer: CountDownTimer = object : CountDownTimer(3000, 1000) {
-        override fun onTick(millisUntilFinished: Long) {
-
-        }
-
-        override fun onFinish() {
-            if (AppPreferences.loginUuid == null) {
-                findNavController().navigate(R.id.loginFragment)
-            } else {
-                findNavController().navigate(R.id.homeFragment)
-            }
-
-        }
-    }.start();
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentSplashBinding.inflate(layoutInflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-        timer.cancel()
-    }
-
-    override fun onResume() {
-        super.onResume()
-        timer.start()
-    }
-
-
-    companion object {
-        const val TAG: String = "SplashFragment"
-    }
-
-}
