+++ app/src/main/res/drawable/ic_circle.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="512dp"
+    android:height="512dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:pathData="M238.5,14c-99.6,7.3 -185.2,75.4 -214.4,170.5 -12.9,42.3 -14.1,88.9 -3.2,131.5 22,85.8 89.1,153.1 174.8,175 28.2,7.2 63.1,9.3 91.8,5.5 53.3,-7 102.6,-31.4 140.5,-69.6 58.5,-59.1 82.6,-143 64.4,-224.8 -6.9,-31 -22,-64 -40.8,-89.6 -38.8,-52.6 -96.6,-87.3 -160.6,-96.4 -16.3,-2.3 -37.6,-3.2 -52.5,-2.1z"
+      android:fillColor="#000000"
+      android:strokeColor="#00000000"/>
+</vector>
