+++ app/src/main/res/drawable/top_radius.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <corners
+        android:topLeftRadius="12dp"
+        android:topRightRadius="12dp"
+        />
+
+    <solid
+        android:color="@color/white"/>
+
+</shape>
